rulebinding cache to be shared for rule id's in ap
