+++ pkg/ruleengine/v1/helpers.go
@@ -20,6 +20,9 @@ import (
 // SensitiveFiles is a list of sensitive files that should not be accessed by the application unexpectedly.
 var SensitiveFiles = []string{
 	"/etc/shadow",
+	"/etc/sudoers",
+	"/etc/ssh/ssh_config",
+	"/etc/ssh/sshd_config",
 }
 
 var (
